moved files around so that the main ones are under "src"
