--- front-end/webpack.config.js
@@ -1,29 +1,0 @@
-var HTMLWebpackPlugin = require('html-webpack-plugin');
-var HTMLWebpackPluginConfig = new HTMLWebpackPlugin({
-    template: __dirname + '/app/index.html',
-    filename: 'index.html',
-    inject: 'body'
-});
-module.exports = {
-    entry: __dirname + '/app/index.js',
-    mode: 'development',
-    performance: {
-        hints: false,
-        maxEntrypointSize: 512000,
-        maxAssetSize: 512000
-    },
-    module: {
-        rules: [
-            {
-                test: /\.js$/,
-                exclude: /node_modules/,
-                loader: 'babel-loader'
-            }
-        ]
-    },
-    output: {
-        filename: 'transformed.js',
-        path: __dirname + '/build'
-    },
-    plugins: [HTMLWebpackPluginConfig]
-};
