+++ src/front-end/README.md
@@ -1,0 +1,11 @@
+# Recipe Cart Frontend Development
+
+## Steps for local React dev
+- Install Node.js (v18)
+- Clone the repo
+- In your local repo directory run:
+    - `npm i`
+    - `npm run build`
+    - `npm run start`
+- Access the website at `http://localhost:8080`
+- See your changes live in action when you edit JavaScript files!
